+++ packages/validation/schemas/login/index.ts
@@ -1,2 +1,2 @@
-export * as code from "./code";
+export * as loginCode from "./loginCode";
 export * as email from "./email";
